Lots of back and forth here
basically:
The asset loader and game renderer are an absolutely mess
Needs to be fixed, not necessary right now but it's pretty awful

+++ core/src/com/jormandr/ui/UIButton.java
@@ -42,8 +42,9 @@ public class UIButton {
 
 	}
 
-	public boolean isMouseOver(float screenX, float screenY) {
-		return (Intersector.isPointInPolygon(verts, 0, 8, screenX, screenY));
+	public boolean isMouseOver(int screenX, int screenY) {
+		
+		return (Intersector.isPointInPolygon(verts, 0, 8, (float)screenX, (float)screenY));
 	}
 
 	public void draw(SpriteBatch batcher) {
